FIX: warnings

+++ Core/Algorithms/mitkImageSliceSelector.cpp
@@ -31,12 +31,12 @@ void mitk::ImageSliceSelector::GenerateOutputInformation()
 
 	output->Initialize(input->GetPixelType(), 2, input->GetDimensions());
 
-  if(m_SliceNr >= input->GetDimension(2))
+  if( (unsigned int)m_SliceNr >= input->GetDimension(2) )
   {
     m_SliceNr = input->GetDimension(2)-1;
   }
 
-  if(m_TimeNr >= input->GetDimension(3))
+  if( (unsigned int)m_TimeNr >= input->GetDimension(3) )
   {
     m_TimeNr = input->GetDimension(3)-1;
   }
